+++ src/Calibre/CustomColumnTypeInteger.php
@@ -156,6 +156,7 @@ class CustomColumnTypeInteger extends CustomColumnType
             $range = $post->min_value . "-" . $post->max_value;
             $params = ['custom' => $this->customId, 'range' => $range, 'db' => $this->databaseId];
             // @todo if we want to use ROUTE_DETAIL we need to add id= here
+            $params['id'] = '0';
             $href = $this->handler::route($routeName, $params);
             array_push($entryArray, new Entry(
                 $range,